refactor(protractor): share the same config across all the Protractor examples
