+++ protractor.conf.js
@@ -1,29 +1,16 @@
-const localChromeConfig = {
-    chromeDriver: require(`chromedriver`).path,
-    directConnect: true,
+const { computeExecutablePath } = require('@puppeteer/browsers')
+const { resolve } = require('path')
 
-    capabilities: {
-        browserName: 'chrome',
-        acceptInsecureCerts : true,
+// Chrome 129 is the last version that correctly supports Selenium 3
+// Chrome 130 and later require Selenium 4 for browser.executeScript to correctly resolve WebElement arguments
+const defaults = {
+    buildId: '129',
+    cacheDir: resolve(__dirname, './browsers'),
+};
 
-        loggingPrefs: {
-            browser: 'INFO',
-        },
-
-        'goog:chromeOptions': {
-            // As of version 75, ChromeDriver is W3C by default, which Protractor does not fully support.
-            w3c: false,
-            excludeSwitches: [ 'enable-automation' ],
-            args: [
-                '--disable-web-security',
-                '--allow-file-access-from-files',
-                '--allow-file-access',
-                '--headless',
-                '--disable-gpu',
-                '--window-size=1024x768',
-            ]
-        },
-    }
+const binaries = {
+    chromedriver: computeExecutablePath({ browser: 'chromedriver', ...defaults }),
+    chrome: computeExecutablePath({ browser: 'chrome', ...defaults }),
 }
 
 module.exports = {
@@ -36,17 +23,44 @@ module.exports = {
 
     allScriptsTimeout: 120_000,
 
+    // specs: [ ],  // set in module-specific protractor.conf.js files
+
     framework: 'mocha',
-
-    // specs: [ ],  // set in module-specific protractor.conf.js
-
     mochaOpts: {
         timeout: 60_000,
+        // ts-node is already loaded by nyc when protractor is executed via npm test
         require: [
             'ts-node/register',
         ],
         reporter: 'spec',
     },
 
-    ... localChromeConfig,
+
+    directConnect: true,
+
+    chromeDriver: binaries.chromedriver,
+
+    capabilities: {
+        browserName: 'chrome',
+        acceptInsecureCerts : true,
+
+        loggingPrefs: {
+            browser: 'INFO',
+        },
+
+        'goog:chromeOptions': {
+            // As of version 75, ChromeDriver is W3C by default, which Protractor does not fully support.
+            w3c: false,
+            binary: binaries.chrome,
+            excludeSwitches: [ 'enable-automation' ],
+            args: [
+                '--disable-web-security',
+                '--allow-file-access-from-files',
+                '--allow-file-access',
+                '--headless',
+                '--disable-gpu',
+                '--window-size=1024x768',
+            ]
+        },
+    },
 };
